rename wmp_move_gen (add _) and add some wmp test in move_gen_test
