+++ test/tsrc/move_gen_test.c
@@ -525,10 +525,14 @@ void top_equity_play_recorder_test(void) {
   config_destroy(config);
 }
 
-void distinct_lexica_test(void) {
-  Config *config =
-      config_create_or_die("set -l1 CSW21 -l2 NWL20 -s1 equity -s2 equity "
-                           "-r1 best -r2 best -numplays 1");
+void distinct_lexica_test(bool wmp1) {
+  Config *config = wmp1
+                       ? config_create_or_die(
+                             "set -l1 CSW21 -l2 NWL20 -s1 equity -s2 equity "
+                             "-r1 best -r2 best -numplays 1 -wmp1 true")
+                       : config_create_or_die(
+                             "set -l1 CSW21 -l2 NWL20 -s1 equity -s2 equity "
+                             "-r1 best -r2 best -numplays 1");
   Game *game = config_game_create(config);
   const LetterDistribution *ld = game_get_ld(game);
   MoveList *move_list = move_list_create(1);
@@ -832,7 +836,8 @@ void test_move_gen(void) {
   exchange_tests();
   equity_test();
   top_equity_play_recorder_test();
-  distinct_lexica_test();
+  distinct_lexica_test(false);
+  distinct_lexica_test(true);
   consistent_tiebreaking_test();
   wordsmog_test();
   movegen_game_update_test();
