+++ test/tsrc/test.c
@@ -42,7 +42,7 @@
 #include "wasm_api_test.h"
 #include "wmp_test.h"
 #include "wmp_maker_test.h"
-#include "wmp_movegen_test.h"
+#include "wmp_move_gen_test.h"
 #include "word_prune_test.h"
 #include "word_test.h"
 
@@ -59,6 +59,9 @@ void run_all(void) {
   test_leave_map();
   test_kwg_alpha();
   test_checkpoint();
+
+  // This creates testdata/lexica/CSW21.wmp that can be used by other tests.
+  test_wmp();
 
   // Now test the rest
   test_bag();
@@ -87,7 +90,6 @@ void run_all(void) {
   test_word_prune();
   test_kwg_maker();
   test_wmp_maker();
-  test_wmp();
   test_wmp_movegen();
   test_file_handler();
   test_leave_list();
